rename region api to regions
