+++ backend/src/api/node_steward_api/mod.rs
@@ -6,5 +6,5 @@ pub fn node_steward_api_router() -> OpenApiRouter {
     OpenApiRouter::new()
         .nest("/this_region_node", routes::this_region_node::router())
         .nest("/local_apps", routes::local_apps::router())
-        .nest("/region", routes::region::router())
+        .nest("/regions", routes::regions::router())
 }
