+++ backend/src/api/node_steward_api/routes/mod.rs
@@ -1,3 +1,3 @@
 pub mod local_apps;
-pub mod region;
+pub mod regions;
 pub mod this_region_node;
