+++ backend/src/api/public_api/routes/mod.rs
@@ -1,8 +1,8 @@
 pub mod local_apps;
 pub mod network;
 pub mod nodes;
-pub mod region;
 pub mod region_apps;
+pub mod regions;
 pub mod stacks;
 pub mod this_p2panda_node;
 pub mod this_region_node;
